[remove server] removed ServerParserBuilder from core/usage.py and moved it to a new file detect_secrets_server/usage.py
